moved default implementations to base feature;
