+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "bytecon"
-version = "0.2.10"
+version = "0.3.0"
 edition = "2021"
 authors = ["Austin Heller"]
 description = "A trait for easily converting to-and-from an array of bytes."
@@ -11,11 +11,11 @@ keywords = ["encoding", "trait", "serialization", "networking", "streaming"]
 categories = ["encoding"]
 
 [features]
-default = []
+base = []
 burn = ["dep:burn"]
 burn_dtype = ["dep:burn"]
 tokio = ["dep:tokio", "dep:tokio-rustls"]
-bincode = ["dep:bincode"]
+bincode = []
 rand = ["dep:rand", "dep:rand_chacha"]
 rustls = ["dep:rustls"]
 bevy = ["dep:bevy", "glam"]
@@ -23,7 +23,7 @@ glam = ["dep:glam"]
 
 [dependencies]
 bevy = { version = "0.15", optional = true }
-bincode = { version = "1", optional = true }
+bincode = { version = "1" }
 burn = { version = "0.14", optional = true }
 glam = { version = "0.29", optional = true }
 paste = { version = "1" }
