+++ src/glam.rs
@@ -2,6 +2,7 @@ use crate::ByteConverter;
 use glam::{Affine2, BVec2, BVec3, BVec3A, BVec4, DAffine2, DAffine3, DMat2, DMat3, DMat4, DQuat, DVec2, DVec3, DVec4, I16Vec2, I16Vec3, I16Vec4, I64Vec2, I64Vec3, I64Vec4, I8Vec2, I8Vec3, I8Vec4, IVec2, IVec3, IVec4, Mat2, Mat3, Mat3A, Mat4, Quat, U16Vec2, U16Vec3, U16Vec4, U64Vec2, U64Vec3, U64Vec4, U8Vec2, U8Vec3, U8Vec4, UVec2, UVec3, UVec4, Vec2, Vec3, Vec3A, Vec4};
 
 impl ByteConverter for Quat {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
@@ -9,6 +10,7 @@ impl ByteConverter for Quat {
         self.w.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self::from_xyzw(
             f32::extract_from_bytes(bytes, index)?,
@@ -20,12 +22,14 @@ impl ByteConverter for Quat {
 }
 
 impl ByteConverter for Vec3 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         self.z.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self::new(
             f32::extract_from_bytes(bytes, index)?,
@@ -36,11 +40,13 @@ impl ByteConverter for Vec3 {
 }
 
 impl ByteConverter for Vec2 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self::new(
             f32::extract_from_bytes(bytes, index)?,
@@ -50,11 +56,13 @@ impl ByteConverter for Vec2 {
 }
 
 impl ByteConverter for Mat2 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x_axis.append_to_bytes(bytes)?;
         self.y_axis.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self::from_cols(
             Vec2::extract_from_bytes(bytes, index)?,
@@ -64,12 +72,14 @@ impl ByteConverter for Mat2 {
 }
 
 impl ByteConverter for Mat3 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x_axis.append_to_bytes(bytes)?;
         self.y_axis.append_to_bytes(bytes)?;
         self.z_axis.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self::from_cols(
             Vec3::extract_from_bytes(bytes, index)?,
@@ -80,11 +90,13 @@ impl ByteConverter for Mat3 {
 }
 
 impl ByteConverter for Affine2 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.matrix2.append_to_bytes(bytes)?;
         self.translation.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             matrix2: Mat2::extract_from_bytes(bytes, index)?,
@@ -94,11 +106,13 @@ impl ByteConverter for Affine2 {
 }
 
 impl ByteConverter for BVec2 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: bool::extract_from_bytes(bytes, index)?,
@@ -108,12 +122,14 @@ impl ByteConverter for BVec2 {
 }
 
 impl ByteConverter for BVec3 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         self.z.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: bool::extract_from_bytes(bytes, index)?,
@@ -124,6 +140,7 @@ impl ByteConverter for BVec3 {
 }
 
 impl ByteConverter for BVec3A {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         let bitmask = self.bitmask();
         let x = (bitmask & (1 << 0)) != 0;
@@ -134,6 +151,7 @@ impl ByteConverter for BVec3A {
         z.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self::new(
             bool::extract_from_bytes(bytes, index)?,
@@ -144,6 +162,7 @@ impl ByteConverter for BVec3A {
 }
 
 impl ByteConverter for BVec4 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
@@ -151,6 +170,7 @@ impl ByteConverter for BVec4 {
         self.w.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: bool::extract_from_bytes(bytes, index)?,
@@ -162,11 +182,13 @@ impl ByteConverter for BVec4 {
 }
 
 impl ByteConverter for DVec2 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: f64::extract_from_bytes(bytes, index)?,
@@ -176,11 +198,13 @@ impl ByteConverter for DVec2 {
 }
 
 impl ByteConverter for DMat2 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x_axis.append_to_bytes(bytes)?;
         self.y_axis.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x_axis: DVec2::extract_from_bytes(bytes, index)?,
@@ -190,11 +214,13 @@ impl ByteConverter for DMat2 {
 }
 
 impl ByteConverter for DAffine2 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.matrix2.append_to_bytes(bytes)?;
         self.translation.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             matrix2: DMat2::extract_from_bytes(bytes, index)?,
@@ -204,12 +230,14 @@ impl ByteConverter for DAffine2 {
 }
 
 impl ByteConverter for DVec3 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         self.z.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: f64::extract_from_bytes(bytes, index)?,
@@ -220,11 +248,13 @@ impl ByteConverter for DVec3 {
 }
 
 impl ByteConverter for DMat3 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x_axis.append_to_bytes(bytes)?;
         self.y_axis.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x_axis: DVec3::extract_from_bytes(bytes, index)?,
@@ -235,11 +265,13 @@ impl ByteConverter for DMat3 {
 }
 
 impl ByteConverter for DAffine3 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.matrix3.append_to_bytes(bytes)?;
         self.translation.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             matrix3: DMat3::extract_from_bytes(bytes, index)?,
@@ -249,6 +281,7 @@ impl ByteConverter for DAffine3 {
 }
 
 impl ByteConverter for DVec4 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
@@ -256,6 +289,7 @@ impl ByteConverter for DVec4 {
         self.w.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: f64::extract_from_bytes(bytes, index)?,
@@ -267,6 +301,7 @@ impl ByteConverter for DVec4 {
 }
 
 impl ByteConverter for DMat4 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x_axis.append_to_bytes(bytes)?;
         self.y_axis.append_to_bytes(bytes)?;
@@ -274,6 +309,7 @@ impl ByteConverter for DMat4 {
         self.w_axis.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x_axis: DVec4::extract_from_bytes(bytes, index)?,
@@ -285,6 +321,7 @@ impl ByteConverter for DMat4 {
 }
 
 impl ByteConverter for DQuat {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
@@ -292,6 +329,7 @@ impl ByteConverter for DQuat {
         self.w.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: f64::extract_from_bytes(bytes, index)?,
@@ -303,11 +341,13 @@ impl ByteConverter for DQuat {
 }
 
 impl ByteConverter for I16Vec2 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: i16::extract_from_bytes(bytes, index)?,
@@ -317,12 +357,14 @@ impl ByteConverter for I16Vec2 {
 }
 
 impl ByteConverter for I16Vec3 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         self.z.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: i16::extract_from_bytes(bytes, index)?,
@@ -333,6 +375,7 @@ impl ByteConverter for I16Vec3 {
 }
 
 impl ByteConverter for I16Vec4 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
@@ -340,6 +383,7 @@ impl ByteConverter for I16Vec4 {
         self.w.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: i16::extract_from_bytes(bytes, index)?,
@@ -351,11 +395,13 @@ impl ByteConverter for I16Vec4 {
 }
 
 impl ByteConverter for I64Vec2 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: i64::extract_from_bytes(bytes, index)?,
@@ -365,12 +411,14 @@ impl ByteConverter for I64Vec2 {
 }
 
 impl ByteConverter for I64Vec3 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         self.z.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: i64::extract_from_bytes(bytes, index)?,
@@ -381,6 +429,7 @@ impl ByteConverter for I64Vec3 {
 }
 
 impl ByteConverter for I64Vec4 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
@@ -388,6 +437,7 @@ impl ByteConverter for I64Vec4 {
         self.w.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: i64::extract_from_bytes(bytes, index)?,
@@ -399,11 +449,13 @@ impl ByteConverter for I64Vec4 {
 }
 
 impl ByteConverter for I8Vec2 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: i8::extract_from_bytes(bytes, index)?,
@@ -413,12 +465,14 @@ impl ByteConverter for I8Vec2 {
 }
 
 impl ByteConverter for I8Vec3 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         self.z.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: i8::extract_from_bytes(bytes, index)?,
@@ -429,6 +483,7 @@ impl ByteConverter for I8Vec3 {
 }
 
 impl ByteConverter for I8Vec4 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
@@ -436,6 +491,7 @@ impl ByteConverter for I8Vec4 {
         self.w.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: i8::extract_from_bytes(bytes, index)?,
@@ -447,11 +503,13 @@ impl ByteConverter for I8Vec4 {
 }
 
 impl ByteConverter for IVec2 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: i32::extract_from_bytes(bytes, index)?,
@@ -461,12 +519,14 @@ impl ByteConverter for IVec2 {
 }
 
 impl ByteConverter for IVec3 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         self.z.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: i32::extract_from_bytes(bytes, index)?,
@@ -477,6 +537,7 @@ impl ByteConverter for IVec3 {
 }
 
 impl ByteConverter for IVec4 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
@@ -484,6 +545,7 @@ impl ByteConverter for IVec4 {
         self.w.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: i32::extract_from_bytes(bytes, index)?,
@@ -495,12 +557,14 @@ impl ByteConverter for IVec4 {
 }
 
 impl ByteConverter for Vec3A {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         self.z.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self::new(
             f32::extract_from_bytes(bytes, index)?,
@@ -511,12 +575,14 @@ impl ByteConverter for Vec3A {
 }
 
 impl ByteConverter for Mat3A {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x_axis.append_to_bytes(bytes)?;
         self.y_axis.append_to_bytes(bytes)?;
         self.z_axis.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x_axis: Vec3A::extract_from_bytes(bytes, index)?,
@@ -527,6 +593,7 @@ impl ByteConverter for Mat3A {
 }
 
 impl ByteConverter for Vec4 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
@@ -534,6 +601,7 @@ impl ByteConverter for Vec4 {
         self.w.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self::new(
             f32::extract_from_bytes(bytes, index)?,
@@ -545,6 +613,7 @@ impl ByteConverter for Vec4 {
 }
 
 impl ByteConverter for Mat4 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x_axis.append_to_bytes(bytes)?;
         self.y_axis.append_to_bytes(bytes)?;
@@ -552,6 +621,7 @@ impl ByteConverter for Mat4 {
         self.w_axis.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x_axis: Vec4::extract_from_bytes(bytes, index)?,
@@ -563,11 +633,13 @@ impl ByteConverter for Mat4 {
 }
 
 impl ByteConverter for U16Vec2 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: u16::extract_from_bytes(bytes, index)?,
@@ -577,12 +649,14 @@ impl ByteConverter for U16Vec2 {
 }
 
 impl ByteConverter for U16Vec3 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         self.z.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: u16::extract_from_bytes(bytes, index)?,
@@ -593,6 +667,7 @@ impl ByteConverter for U16Vec3 {
 }
 
 impl ByteConverter for U16Vec4 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
@@ -600,6 +675,7 @@ impl ByteConverter for U16Vec4 {
         self.w.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: u16::extract_from_bytes(bytes, index)?,
@@ -611,11 +687,13 @@ impl ByteConverter for U16Vec4 {
 }
 
 impl ByteConverter for U64Vec2 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: u64::extract_from_bytes(bytes, index)?,
@@ -625,12 +703,14 @@ impl ByteConverter for U64Vec2 {
 }
 
 impl ByteConverter for U64Vec3 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         self.z.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: u64::extract_from_bytes(bytes, index)?,
@@ -641,6 +721,7 @@ impl ByteConverter for U64Vec3 {
 }
 
 impl ByteConverter for U64Vec4 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
@@ -648,6 +729,7 @@ impl ByteConverter for U64Vec4 {
         self.w.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: u64::extract_from_bytes(bytes, index)?,
@@ -659,11 +741,13 @@ impl ByteConverter for U64Vec4 {
 }
 
 impl ByteConverter for U8Vec2 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: u8::extract_from_bytes(bytes, index)?,
@@ -673,12 +757,14 @@ impl ByteConverter for U8Vec2 {
 }
 
 impl ByteConverter for U8Vec3 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         self.z.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: u8::extract_from_bytes(bytes, index)?,
@@ -689,6 +775,7 @@ impl ByteConverter for U8Vec3 {
 }
 
 impl ByteConverter for U8Vec4 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
@@ -696,6 +783,7 @@ impl ByteConverter for U8Vec4 {
         self.w.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: u8::extract_from_bytes(bytes, index)?,
@@ -707,11 +795,13 @@ impl ByteConverter for U8Vec4 {
 }
 
 impl ByteConverter for UVec2 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: u32::extract_from_bytes(bytes, index)?,
@@ -721,12 +811,14 @@ impl ByteConverter for UVec2 {
 }
 
 impl ByteConverter for UVec3 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
         self.z.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: u32::extract_from_bytes(bytes, index)?,
@@ -737,6 +829,7 @@ impl ByteConverter for UVec3 {
 }
 
 impl ByteConverter for UVec4 {
+    #[inline(always)]
     fn append_to_bytes(&self, bytes: &mut Vec<u8>) -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
         self.x.append_to_bytes(bytes)?;
         self.y.append_to_bytes(bytes)?;
@@ -744,6 +837,7 @@ impl ByteConverter for UVec4 {
         self.w.append_to_bytes(bytes)?;
         Ok(())
     }
+    #[inline(always)]
     fn extract_from_bytes(bytes: &Vec<u8>, index: &mut usize) -> Result<Self, Box<dyn std::error::Error + Send + Sync + 'static>> where Self: Sized {
         Ok(Self {
             x: u32::extract_from_bytes(bytes, index)?,
