+++ test.sh
@@ -1,3 +1,3 @@
 export RUST_BACKTRACE=1
-cargo test --all-features -- --show-output
+cargo test --all-features --release -- --show-output
 read -n 1 -s
